Building AI Powered Solution for Assisting Visually Impaired Individuals/requirements.txt
streamlit 
google-generativeai 
pytesseract 
gtts 
pillow
